+++ evalml/automl/auto_regression_search.py
@@ -19,7 +19,6 @@ class AutoRegressionSearch(AutoSearchBase):
                  allowed_model_families=None,
                  cv=None,
                  tuner=None,
-                 detect_label_leakage=True,
                  start_iteration_callback=None,
                  add_result_callback=None,
                  additional_objectives=None,
@@ -52,9 +51,6 @@ class AutoRegressionSearch(AutoSearchBase):
             cv: cross validation method to use. By default StratifiedKFold
 
             tuner: the tuner class to use. Defaults to scikit-optimize tuner
-
-            detect_label_leakage (bool): If True, check input features for label leakage and
-                warn if found. Defaults to true.
 
             start_iteration_callback (callable): function called before each pipeline training iteration.
                 Passed two parameters: pipeline_class, parameters.
@@ -91,7 +87,6 @@ class AutoRegressionSearch(AutoSearchBase):
             tolerance=tolerance,
             allowed_model_families=allowed_model_families,
             problem_type=problem_type,
-            detect_label_leakage=detect_label_leakage,
             start_iteration_callback=start_iteration_callback,
             add_result_callback=add_result_callback,
             additional_objectives=additional_objectives,