+++ evalml/pipelines/components/estimators/classifiers/__init__.py
@@ -3,3 +3,4 @@ from .logistic_regression import LogisticRegressionClassifier
 from .rf_classifier import RandomForestClassifier
 from .xgboost_classifier import XGBoostClassifier
 from .catboost_classifier import CatBoostClassifier
+from .baseline_classifier import BaselineClassifier