+++ evalml/tests/pipeline_tests/regression_pipeline_tests/test_baseline_regression.py
@@ -1,0 +1,31 @@
+import numpy as np
+
+from evalml.pipelines import BaselineRegressionPipeline
+
+
+def test_baseline_mean(X_y_reg):
+    X, y = X_y_reg
+    mean = y.mean()
+    parameters = {
+        "Baseline Regressor": {
+            "strategy": "mean"
+        }
+    }
+    clf = BaselineRegressionPipeline(parameters=parameters)
+    clf.fit(X, y)
+    np.testing.assert_allclose(clf.predict(X), np.array([mean] * len(X)))
+    np.testing.assert_allclose(clf.feature_importances.iloc[:, 1], np.array([0.0] * X.shape[1]))
+
+
+def test_baseline_median(X_y_reg):
+    X, y = X_y_reg
+    median = np.median(y)
+    parameters = {
+        "Baseline Regressor": {
+            "strategy": "median"
+        }
+    }
+    clf = BaselineRegressionPipeline(parameters=parameters)
+    clf.fit(X, y)
+    np.testing.assert_allclose(clf.predict(X), np.array([median] * len(X)))
+    np.testing.assert_allclose(clf.feature_importances.iloc[:, 1], np.array([0.0] * X.shape[1]))